requirement.txt
asgiref==3.7.1
backports.zoneinfo==0.2.1
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.2
